+++ src/PointInTime/Month.php
@@ -3,41 +3,32 @@ declare(strict_types = 1);
 
 namespace Innmind\TimeContinuum\PointInTime;
 
+use Innmind\TimeContinuum\Calendar;
+
 /**
  * @psalm-immutable
  */
 final class Month
 {
-    /** @var int<1, 12> */
-    private int $month;
+    private Calendar\Month $month;
     /** @var int<28, 31> */
     private int $days;
-    /** @var 'January'|'February'|'March'|'April'|'May'|'June'|'July'|'August'|'September'|'October'|'November'|'December' */
-    private string $string;
 
-    /**
-     * @param int<1, 12> $month
-     */
-    private function __construct(Year $year, int $month)
+    private function __construct(Year $year, Calendar\Month $month)
     {
-        // todo transform into an enum
         $this->month = $month;
         /** @var int<28, 31> */
         $this->days = (int) \date(
             't',
-            $time = \mktime(0, 0, 0, $month, 1, $year->toInt()),
+            \mktime(0, 0, 0, $month->toInt(), 1, $year->toInt()),
         );
-        /** @var 'January'|'February'|'March'|'April'|'May'|'June'|'July'|'August'|'September'|'October'|'November'|'December' */
-        $this->string = \date('F', $time);
     }
 
     /**
      * @psalm-pure
      * @internal
-     *
-     * @param int<1, 12> $month
      */
-    public static function of(Year $year, int $month): self
+    public static function of(Year $year, Calendar\Month $month): self
     {
         return new self($year, $month);
     }
@@ -50,19 +41,8 @@ final class Month
         return $this->days;
     }
 
-    /**
-     * @return int<1, 12>
-     */
-    public function toInt(): int
+    public function calendar(): Calendar\Month
     {
         return $this->month;
-    }
-
-    /**
-     * @return 'January'|'February'|'March'|'April'|'May'|'June'|'July'|'August'|'September'|'October'|'November'|'December'
-     */
-    public function toString(): string
-    {
-        return $this->string;
     }
 }