+++ tests/NowTest.php
@@ -39,7 +39,7 @@ class NowTest extends TestCase
         $this->assertInstanceOf(Millisecond::class, $point->millisecond());
         $this->assertInstanceOf(Offset::class, $point->offset());
         $this->assertSame((int) \date('Y', $timestamp), $point->year()->toInt());
-        $this->assertSame((int) \date('m', $timestamp), $point->month()->toInt());
+        $this->assertSame((int) \date('m', $timestamp), $point->month()->calendar()->toInt());
         $this->assertSame((int) \date('d', $timestamp), $point->day()->toInt());
         $this->assertSame((int) \date('H', $timestamp), $point->hour()->toInt());
         $this->assertSame((int) \date('i', $timestamp), $point->minute()->toInt());
@@ -83,7 +83,7 @@ class NowTest extends TestCase
         $this->assertNotSame($point->millisecond(), $point2->millisecond());
         $this->assertSame($point->milliseconds(), $point2->milliseconds());
         $this->assertSame((int) $now->format('Y'), $point2->year()->toInt());
-        $this->assertSame((int) $now->format('m'), $point2->month()->toInt());
+        $this->assertSame((int) $now->format('m'), $point2->month()->calendar()->toInt());
         $this->assertSame((int) $now->format('d'), $point2->day()->toInt());
         $this->assertSame((int) $now->format('H'), $point2->hour()->toInt());
         $this->assertSame((int) $now->format('i'), $point2->minute()->toInt());