+++ tests/PointInTime/DayTest.php
@@ -7,6 +7,7 @@ use Innmind\TimeContinuum\{
     PointInTime\Day,
     PointInTime\Month,
     PointInTime\Year,
+    Calendar,
 };
 use PHPUnit\Framework\TestCase;
 
@@ -14,7 +15,14 @@ class DayTest extends TestCase
 {
     public function testInterface()
     {
-        $day = Day::of($year = Year::of(2016), Month::of($year, 10), 5);
+        $day = Day::of(
+            $year = Year::of(2016),
+            Month::of(
+                $year,
+                Calendar\Month::of(10),
+            ),
+            5,
+        );
 
         $this->assertSame(3, $day->weekNumber());
         $this->assertSame(278, $day->ofYear());