+++ tests/PointInTime/MonthTest.php
@@ -6,6 +6,7 @@ namespace Tests\Innmind\TimeContinuum\PointInTime;
 use Innmind\TimeContinuum\{
     PointInTime\Month,
     PointInTime\Year,
+    Calendar,
 };
 use PHPUnit\Framework\TestCase;
 
@@ -13,10 +14,10 @@ class MonthTest extends TestCase
 {
     public function testInterface()
     {
-        $month = Month::of(Year::of(2016), 10);
+        $month = Month::of(Year::of(2016), Calendar\Month::of(10));
 
         $this->assertSame(31, $month->numberOfDays());
-        $this->assertSame(10, $month->toInt());
-        $this->assertSame('October', $month->toString());
+        $this->assertSame(10, $month->calendar()->toInt());
+        $this->assertSame('october', $month->calendar()->name);
     }
 }