+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Added
 
 - `Innmind\TimeContinuum\Clock::switch()`
+- `Innmind\TimeContinuum\Calendar\Month`
+- `Innmind\TimeContinuum\PointInTime\Month::calendar()`
 
 ### Changed
 
@@ -39,6 +41,8 @@
 - Earth timezones
 - `Innmind\TimeContinuum\Timezone::daylightSavingTimeApplied()`
 - `Innmind\TimeContinuum\Exception`
+- `Innmind\TimeContinuum\PointInTime\Month::toInt()`
+- `Innmind\TimeContinuum\PointInTime\Month::toString()`
 
 ## 3.4.1 - 2023-09-17